weather_api/weather/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
alembic
python-jose[cryptography]
